Project-2
Netflix Homepage Clone: A redesigned replica of the classic Netflix homepage. This project, built using HTML and CSS, duplicates the layout and appearance of the Netflix website, providing a streamlined, visually attractive interface. Dive into the code to learn about the structure and styles of this front-end project, which is great for practising and learning web development skills.